requirements.txt
# ML & Data
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0
pyarrow>=14.0.0

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# LLM & Agent
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-huggingface>=1.0.0
huggingface-hub>=0.25.0
langgraph>=0.2.0
google-generativeai>=0.8.0

# Frontend
streamlit>=1.29.0

# Visualization & EDA
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.25.0
